Add support to populate and query a database.
This patch proposes initial work to move from a flat files
backend to a database.

Flask-SQLAlchemy and SQLAlchemy are introduced as dependencies
to interface with an rdbms. Flask-SQLAlchemy gives us commodity
methods to setup configuration and request management boilerplate.

The database connection can be configured by setting an appropriate
URI via the SQLALCHEMY_DATABASE_URI configuration option. The
default backend, for development purposes, is in-memory sqlite.

When the service is launched with --resourcedir
a database is created and populated from fixtures at startup
(used for development).

The database is queried for user data (user_text) at request time.
Once data for a user is loaded, three dataset globals are populated.
The data access pattern has been left unchanged - except for some
changes to time fields, which are now stored as Python datetime
objects.

This patch has been tested by providing fixtures and manually
verifying consistency in the output produced by the
endpoint when querying a database vs. loading state from flat files.

Change-Id: I856d553b8f1a2d407fe5d49cea5076c45fc065c1