Add a data ingestion script.
This patch adds a script `migration/ingestion.py` to
load data from the raw tsv datasets into
a database. It extracts and generalises logic from the
load_data() method in wsgi.py.

The script re-uses models and session managment
developed for the service, hence the flask deps.
It's SQLAlchemy under the hood, so the flask machinery
can be factored out if necessary.

Two new fields have been added to the models:
- dataset_id
- insertion_time

This is metadata not used by the service, but
provided for bookeeping and validation/troubleshooting.

An experimental `mysql_lock` context is provided
as a notification mechanism that a batch update
is in progress. The app service should be updated
to test for a lock before answerying a request.
This will be subject of a dedicated CR.

Change-Id: I1e6139ad6d0d83ffade65cb04ff970dc530de906