Abort /similaruser call if db refresh in progress.
When querying user data we should verify that
the database is not being refreshed, and abort
the request if so. The caller will be notfied
with a 403 HTTP status code.

We assume that refreshes are sporadic events, orchestrated by a
separate process. The refresh process won't be notfied of
read attempts. Consistency of the result set is
only eventually guaranteed.
A database refresh can potentially start while executing
the query block.

A new /database/refresh endpoint has been added, that
returns wether a refresh is in progress.

Change-Id: Id1440e8095a7dfa41ed91a756b11558126bc31dc