Throttle database writes.
This patch adds an option to delay a commit
by a number of ms.
This simple models tries to limit insertion
by a given number of rows (batch size) per time window
(the delay duration).

Beacuse of changes in main, app instantiation
had to be refactored to use the create_app()
factory method.

Change-Id: I1b8476eee63f15454dec27f98e7a7de00dc0528a