Allow requests to be made to an alternative endpoint for mwapi
Currently we request to the MW API using
https://lang.wikipedia.org. Unfortunately due to our firewall
configuration in k8s and more generally how internal requests are made
to appservers, we can't follow this pattern. In order to run in k8s,
we will need to make requests to the API LVS services with a Host
header set to the sites we require. Currently the mwapi library does
not support this behaviour, but we can hack around it by setting out
host header in the session object once created by the mwapi Session
constructor. In future it would be best to patch the mwapi library to
get around this but in the interests of getting the service shipped,
hack around this restriction for now.

Also do a quick once-over with `black` to change some quotes and
indentation.

Bug: T268837
Change-Id: Ia9594a5507a2e87063c78757cab356f13c48d6fc