Handle json encoding of binary objects.
This patch adds a JSONEncoder that is able to
handle binary objects. For instance, dictionaries
with binary values.

To be consistent with mediawiki, we are currently
storing user names as varbinary fields.

As a consequence, binary objects might need to be
returned in the get_similar_users json response.

The create_app() factory method registers the
custom serializer, which is picked up by
Flask's jsonify(). As a result,
we don't need to manually decode objects
in wsgi.py anymore.

Change-Id: Ib67808a4559ec77a46f4a0aae8525a93b34b3939